Add files via upload
Second milestone project is a Black Jack game.  it will be single player with betting using OOP.  this should be similar to the War game but with hit or stay and how much to bet.  The trick I need to figure out is how to make the Ace both 1 and 11.  I'm not yet finish with this but I made some progress and have completed the betting class and have the Card and Deck class completed.  I'm thinking I'm going to put the dealer and player's hand in the Player class and then figure out the game logic probably Monday or Tuesday.