Add files via upload
I have been asked to design a final capstone project for my first Bootcamp.  The only requirement is that it is supposed to take at least a week to complete.  My plan, being that my primary focus is data science, is to take a large volume of homogeneous paper documents, in this case, pay stubs, scan them to PDFs run them through an OCR, pull the key data points, and output them into a 2d array.  I am thinking that I will then compare the growth in salary over the course of 3 yrs with the change in the stock market using SPY as a gauge.

For my first day, I have scanned a single document converted it to OCR, then converted that to a list, and pulled the key data points out.  I will need to take at least a full day to scan all of these documents into PDF as there are well over 100 of them.  I should be able to use an enumerate or perhaps a for loop to go through the process I figured out today.  I anticipate being able to pull the key historic data for SPY from my broker without much problem.  I can then use NumPy and Pandas to synthesize the data from there.