Add files via upload
learned a lot about Linux command line and python with this.  didn't get it to quite work the way I wanted it to but I plan to come back to this later and make a second attempt.  I have overcome the OCR problem by converting all files at the command line in Linux, still in one large batch (details within).  I think found that because of the nature of OCR/document imperfections it was unlikely items would always be the same place in a list.  I switched gears and am leaving this with me needing to create some combination of regex and conditional statements that would yield more reliable results.